Automatically applies the DataConstraint when adding an XMLInputSplitter to an text/xml type input
git-svn-id: http://taverna.googlecode.com/svn@1547 bf327186-88b3-11dd-a302-d386e5130c1c

+++ trunk/taverna1.0/src/org/embl/ebi/escience/scuflworkers/java/XMLInputSplitter.java
@@ -161,7 +161,7 @@ public class XMLInputSplitter implements LocalWorker, XMLExtensible {
 					types[i] = String[].class;
 
 				} else if ("double".equals(s)) {
-					types[i] = Integer.TYPE;
+					types[i] = Double.TYPE;
 				} else if ("float".equals(s)) {
 					types[i] = Float.TYPE;
 				} else if ("int".equals(s)) {
