+++ trunk/taverna1.0/src/org/embl/ebi/escience/scuflworkers/wsdl/WSDLBasedProcessor.java
@@ -172,7 +172,7 @@ public class WSDLBasedProcessor extends Processor implements java.io.Serializabl
 					types[i] = String[].class;					
 					
 				} else if ("double".equals(s)) {
-					types[i] = Integer.TYPE;
+					types[i] = Double.TYPE;
 				} else if ("float".equals(s)) {
 					types[i] = Float.TYPE;
 				} else if ("int".equals(s)) {
